Circular dependency solved and notification woprking
