+++ botProject/src/action.js
@@ -4,7 +4,6 @@ const configure = require('./model/config');
 const axios = require('axios');
 const debug = require('debug')('botProject:src/slack_bot');
 const qs = require('querystring');
-const UserModel = require('./model/user');
 
 var Botkit = require('botkit');
 
@@ -49,65 +48,5 @@ module.exports = {
       debug('sendConfirmation error: %o', err);
       console.error(err);
     });
-  },
-  send_tutor_notification: function(user_id, tutor_id, date, day, from, to){
-  // send_tutor_notification(){
-    // var student_name = "Rikne";
-    // var student_phone = "1919191991";
-    // var student_email = "rshah9@ncsu.edu";
-    // var tutor_id = "U93ELGP8S"
-    // var user_id = "U8XDVJD26";
-    // var from = "csdcd";
-    // var to = "cdscds";
-    // var date = "Cdcdsdc";
-    // var day = "cdcdsd";
-    usermodel.get_user_info(user_id, function(user){
-      console.log(user);
-      var student_name = user.name;
-      var student_email = user.email;
-      var student_phone = user.phone;
-      module.exports.send_message(tutor_id,"Hello There",[
-              {
-                  "fields":
-                  [
-                   {
-                      "title": 'Student Name',
-                      "value": student_name,
-                      "short":true,
-                    },
-                    {
-                      "title": 'Email',
-                      "value": student_email,
-                      "short":true,
-                    },
-                    {
-                      "title": 'Phone',
-                      "value": student_phone,
-                      "short":true,
-                    },
-                    {
-                      "title": 'Date',
-                      "value": date,
-                      "short":true,
-                    },
-                    {
-                      "title": 'Day',
-                      "value": day,
-                      "short":true,
-                    },
-                    {
-                      "title": 'From',
-                      "value": from,
-                      "short":true,
-                    },
-                    {
-                      "title": 'To',
-                      "value": to,
-                      "short":true,
-                    }
-                  ]
-              }
-          ]);
-    }); // End of query
-  }// End of function
+  }
 }//end of module
